+++ ea-factories/src/index.ts
@@ -1,0 +1,2 @@
+export * as balance from './factories/balance'
+export * as price from './factories/price'