+++ iex-cloud/src/endpoint/index.ts
@@ -1,0 +1,2 @@
+export * as stock from './stock'
+export * as crypto from './crypto'