+++ iex-cloud/src/endpoint/stock.ts
@@ -1,0 +1,36 @@
+import { Requester, Validator } from '@chainlink/external-adapter'
+import { ExecuteWithConfig, Config } from '@chainlink/types'
+
+export const NAME = 'stock'
+
+const customParams = {
+  base: ['base', 'from', 'coin', 'asset', 'symbol'],
+}
+
+export const execute: ExecuteWithConfig<Config> = async (request, config) => {
+  const validator = new Validator(request, customParams)
+  if (validator.error) throw validator.error
+
+  const jobRunID = validator.validated.id
+  const base = validator.validated.data.base
+  const url = `stock/${base.toUpperCase()}/quote`
+
+  const params = {
+    token: config.apiKey,
+  }
+
+  const reqConfig = {
+    ...config.api,
+    params,
+    url,
+  }
+
+  const response = await Requester.request(reqConfig)
+  const result = Requester.validateResultNumber(response.data, ['latestPrice'])
+
+  return Requester.success(jobRunID, {
+    data: { ...response.data, result },
+    result,
+    status: 200,
+  })
+}