+++ CHANGELOG.md
@@ -17,13 +17,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `finage` to get Financial data from finage.co.uk
   - `coincodex` to get crypto prices from CoinCodex
   - `coinranking` to get crypto prices from Coinranking
+  - `crypto-volatility-index` to calculate the CVI (Crypto volatility index)
+  - `btc.com` to get on-chain balances from BTC.com
+  - `sochain` to get on-chain balances from SoChain
   - `dns-query` to query DNS over HTTPS
   - `dns-record-check` to check whether some record provided exists on DNS
+  - `paxos` to get Paxos asset supply attestations
   - `outlier-detection`: composite adapter to check for outlier values between multiple sets of data providers
+  - `dydx-stark` to sign the input price data with your private STARK key, and send it to the destination endpoint.
+  - `apy-finance` to calculate the TVL in APY Finance
+  - `token-allocation` composite adapter to calculate the total price from a set of tokens
+  - `bitcoin-json-rpc`: composite adapter for querying bitcoin blockchain stats(difficulty, height) according to the existing convention
+  - `iex-cloud` to get stock and crypto market data from IEX Cloud
+  - `cfbenchmarks` to get crypto benchmarks and indices
 - Added support for metadata in requests. This gives adapters access to the FM on-chain round state.
 - Moves re-usable test behaviors & testing utils to a new package - `@chainlink/adapter-test-helpers`
 - Added support for using query string parameters as input to adapters.
+- Add a package re-usable adapter factory implementations to - `@chainlink/ea-factories`
+
 - Added support for using multiple API Keys per-adapter
+
+### Fixed
+
+- CMC adapter price query fixed for currencies where multiple currencies have the same symbol
 
 ### Changed
 
@@ -31,7 +47,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Market closure is now a composite adapter, and now follows the build instructions for composite adapters. Market
   now also supports using metadata.
 - Coinlore now accepts an optional environment variable to set the default API endpoint
-- migrates @chainlink/ea-bootstrap and @chainlink/external-adapter packages to TS.
+- Migrates @chainlink/ea-bootstrap and @chainlink/external-adapter packages to TS.
+- Updates the following adapters with a second endpoint for getting an account's on-chain balance
+  - `amberdata`
+  - `blockchair`
+  - `cryptoapis`
 - dxFeed now uses OTC feeds for FTSE and N225, rather than licensed data feeds
 - migrates `coinmarketcap` to TS